Introduce ParticleInfo
Add a type field to ReferenceParticle (base class for all bunches). This
points to a ParticleInfo class that contains the mass and charge of the
particle species.

ParticleInfoDB contains some pre defined particles.

Classes such as ProtonBunch can set type, rather than using virtual
functions to return properties.

GetParticleLifetime() and IsStable() are removed as they are not used
anywhere.